Move coordinator to separate file
